+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### FEATURES
+- New flag `--persist-hostname` which will update the `/etc/hostname` file value with the generated hostname
+- New flag `--persist-hosts` which will set the generated hostname as a host pointing to the loopback
+
 ### ENHANCEMENTS
 - Makefile improvements
 - Updated dependencies to their latest versions